refactor scroll fade logic into a reusable hook
