+++ packages/synapse-react-client/src/components/SynapseHomepageV2/SynapseInActionItem.tsx
@@ -7,10 +7,11 @@ import {
   useMediaQuery,
   SxProps,
 } from '@mui/material'
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import ImageFromSynapseTable from '../ImageFromSynapseTable'
 import { EastTwoTone } from '@mui/icons-material'
 import { darkTextColor, homepageBodyText } from './SynapseHomepageV2'
+import { useScrollFadeTransition } from 'src/utils/hooks/useScrollFadeTransition'
 
 export type SynapseInActionItemProps = {
   tableId: string
@@ -34,26 +35,6 @@ const mobileViewSxProps: SxProps = {
   textAlign: 'center',
 }
 
-const calculateOpacity = (rect: DOMRect): number => {
-  const viewportHeight = window.innerHeight
-  const elementCenterY = rect.top + rect.height / 2
-  const middleRangeStart = viewportHeight * 0.3
-  const middleRangeEnd = viewportHeight * 0.7
-
-  if (elementCenterY >= middleRangeStart && elementCenterY <= middleRangeEnd) {
-    // center of element is within the viewport bounds that I want to show full opacity
-    return 1
-  } else {
-    //otherwise, subtract the % (of the center of the element to one of the boundaries)
-    const distanceToMiddle = Math.min(
-      Math.abs(elementCenterY - middleRangeStart),
-      Math.abs(elementCenterY - middleRangeEnd),
-    )
-    const maxDistance = viewportHeight / 4
-    return Math.max(0, 1 - distanceToMiddle / maxDistance)
-  }
-}
-
 export const SynapseInActionItem: React.FunctionComponent<
   SynapseInActionItemProps
 > = ({
@@ -70,23 +51,8 @@ export const SynapseInActionItem: React.FunctionComponent<
   const theme = useTheme()
   const isMobileView = useMediaQuery(theme.breakpoints.down('sm'))
 
-  const ref = useRef(null)
   // opacity of the desktop image
-  const [opacity, setOpacity] = useState(1)
-
-  // listen to the scroll event, and calculate the opacity based on where the ref element is in the current viewport
-  useEffect(() => {
-    const handleScroll = () => {
-      const rect = (ref.current as any).getBoundingClientRect()
-      setOpacity(calculateOpacity(rect))
-    }
-
-    if (ref) {
-      window.addEventListener('scroll', handleScroll)
-      handleScroll()
-    }
-    return () => window.removeEventListener('scroll', handleScroll)
-  }, [ref])
+  const { ref, opacity } = useScrollFadeTransition()
 
   return (
     <Box
